fix: dramatically improve fetching speed
This commit fixes the issue when fetching
running foreground applications would take up to 10-15
seconds. The issue occurred because of the additional
overhead created by the osascript shell command, which
would spin up an expensive interpretator layer.

This commit replaces the use of the complex
payload used when calling osascript with call
to "lsappinfo" command (undocumented), which
provides a much quicker way to access a list
of running applications. The result of the shell
command is later parsed and piped into the
next program operation.